// QuantumMind-backend/src/bots/enums/node-type.enum.ts
export enum NodeTypeEnum {
  START_NODE = 'START_NODE',
  USER_INPUT = 'USER_INPUT',
  BOT_RESPONSE = 'BOT_RESPONSE',
  TICKET = 'TICKET',
  SET_ATTRIBUTES = 'SET_ATTRIBUTES',
  CLOSE_CHAT = 'CLOSE_CHAT',
  FILE_ATTACHMENT = 'FILE_ATTACHMENT',
  QUESTIONS = 'QUESTIONS',
  FAQ = 'FAQ',
  WEBHOOK = 'WEBHOOK',
  ADD_TO_SEGMENT = 'ADD_TO_SEGMENT',
  REMOVE_FROM_SEGMENT = 'REMOVE_FROM_SEGMENT',
  TRANSFER_TO_AGENT = 'TRANSFER_TO_AGENT',
  GO_TO_STEP = 'GO_TO_STEP',
  FALL_BACK = 'FALL_BACK',
  SUCCESS = 'SUCCESS',
  FAILURE = 'FAILURE',
}